requirements.txt
aiofiles==0.4.0
beautifulsoup4==4.8.2
blis==0.4.1
Bottleneck==1.3.2
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cycler==0.10.0
cymem==2.0.3
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastai==1.0.60
fastprogress==0.2.2
h11==0.9.0
httptools==0.1.1
idna==2.9
importlib-metadata==1.5.0
imutils==0.5.3
kiwisolver==1.1.0
matplotlib==3.2.0
murmurhash==1.0.2
numexpr==2.7.1
numpy==1.18.1
nvidia-ml-py3==7.352.0
opencv-python==3.4.2.16
opencv-python-headless==4.2.0.32
packaging==20.3
pandas==1.0.1
Pillow==7.0.0
plac==1.1.3
preshed==3.0.2
pynvx==1.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0
spacy==2.2.4
srsly==1.0.2
starlette==0.13.2
thinc==7.4.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.43.0
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
wasabi==0.6.0
websockets==8.1
zipp==3.1.0
